+++ core/realtime-core/src/test/kotlin/org/http4k/filter/CorsAndRebindProtectionTest.kt
@@ -19,7 +19,6 @@ import org.http4k.sse.SseHandler
 import org.http4k.sse.SseResponse
 import org.http4k.websocket.Websocket
 import org.http4k.websocket.WsHandler
-import org.http4k.websocket.WsMessage
 import org.http4k.websocket.WsResponse
 import org.http4k.websocket.WsStatus
 import org.junit.jupiter.api.Test